% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_se_1}
\alias{campp2_brca_se_1}
\title{campp2_brca_se_1 - example data}
\format{
A SummarizedExperiment with 10000 rows (genes) and 8 columns (samples)
}
\usage{
data(campp2_brca_se_1)
}
\value{
A SummarizedExperiment with 10000 rows (genes) and 8 columns (samples)
}
\description{
A campp2_brca_se_1 is a SummarizedExperiment created based on
campp2_brca_1 and campp2_brca_1_meta by running:
rownames(campp2_brca_1_meta)<-campp2_brca_1_meta$ID
campp2_brca_se_1 <- CreateSE(campp2_brca_1, campp2_brca_1_meta)
}
\keyword{datasets}
