% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_se_1_replacedNAs}
\alias{campp2_brca_se_1_replacedNAs}
\alias{campp2_brca_1_replacedNAs}
\title{campp2_brca_se_1_replacedNAs - example data}
\format{
A SummarizedExperiment object with 9993 genes and 8 samples.
}
\usage{
data(campp2_brca_se_1_replacedNAs)
}
\value{
A SummarizedExperiment object with 9993 genes and 8 samples.
}
\description{
A dataset with replaced NA values. This dataset is based on
"campp2_brca_se_1_NAs" dataset. For more details, see the vignette.
}
\keyword{datasets}
