% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campp2_brca_se_2}
\alias{campp2_brca_se_2}
\title{campp2_brca_se_2 - example data}
\format{
A SummarizedExperiment with 10000 rows (genes) and 8 columns (samples)
}
\usage{
data(campp2_brca_se_2)
}
\value{
A SummarizedExperiment with 10000 rows (genes) and 8 columns (samples)
}
\description{
A campp2_brca_se_2 is a SummarizedExperiment created based on
campp2_brca_2 and campp2_brca_2_meta by running:
rownames(campp2_brca_2_meta)<-campp2_brca_2_meta$ID
campp2_brca_se_2 <- create_SE(campp2_brca_2, campp2_brca_2_meta)
}
\keyword{datasets}
