% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSE.R
\name{CreateSE}
\alias{CreateSE}
\title{Create a SummarizedExperiment object}
\usage{
CreateSE(feature_counts, sample_metadata)
}
\arguments{
\item{feature_counts}{A data frame with feature counts. Each column represents a sample,
and each row represents a feature (eg. gene) Column names must match the row names in sample metadata.}

\item{sample_metadata}{A data frame with sample metadata. Each row represents a sample,
and each column represents a metadata category. Row names must match the column names in feature counts}
}
\value{
A SummarizedExperiment object containing the given feature_counts and sample_metadata.
}
\description{
This function creates a SummarizedExperiment object from given feature counts and sample metadata.
It is crucial to ensure that the row names and column names match between the feature counts and sample metadata.
If they do not match, the function will stop and throw an error.
}
\examples{
rownames(campp2_brca_1_meta) <- campp2_brca_1_meta$IDs
campp2_brca_se_1 <- CreateSE(campp2_brca_1, campp2_brca_1_meta)
}
\author{
Nikola Tom
}
