% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_NAs.R
\name{ReplaceNAs}
\alias{ReplaceNAs}
\title{Replace "NA" values}
\usage{
ReplaceNAs(SE, pct.NA.row = 70, pct.NA.column = 80)
}
\arguments{
\item{SE}{a SummarizedExperiment object}

\item{pct.NA.row}{a number defining maximal percentage of NA values present
in a feature (e.g. a percentage of samples having "NA" in a given gene).
Features (rows) with a higher percentage of NA values will be removed (70 by
default).}

\item{pct.NA.column}{a number defining maximal percentage of NA values
present in a sample (e.g. a percentage of genes having NA in one sample).
Samples (columns) with a higher percentage of NA values will be removed
(80 by default).}
}
\value{
Data frame with replaced "NA" values
}
\description{
In the first step, percentages of NAs in rows (features) and
columns (samples) are calculated. Next, features with high
percentage of NAs are removed (by default, features having more than 70
percent). Next, columns (samples) having high percentage of NAs are removed),
80% by default.
Remaining NA values are replaced by imputed values generated by “local least
squares (llsImpute)“. If the first round of missing values imputation failed
or resulted in negative values, an additional missing value imputation using
impute.knn will be performed (rows with more than 50 percent of missing
values will be imputed using the overall mean per sample).
}
\examples{
{
#get a number of genes in the dataset.
n<-nrow(assay(campp2_brca_se_1))

# For testing of this function, 1000 genes is assigned with NA value. NA values are introduced randomly into each sample and all the 10000 genes (n=10000) are considered for the introduction of NA values.
assay(campp2_brca_se_1, withDimnames=FALSE)<-apply (assay(campp2_brca_se_1), 2, function(x) {x[sample( c(1:n), floor(n/10))] <- NA; x} )

# create an object with replaced NAs; features having "NA" in more than 70\% of the samples will be removed; samples having "NA" in more than 80\% of the genes will be removed
campp2_brca_se_1_replacedNAs<-ReplaceNAs(SE=campp2_brca_se_1, pct.NA.row=70, pct.NA.column=80)
}
}
